+++ CHANGELOG.md
@@ -12,7 +12,13 @@
 
 # 11.0.0-alpha.5 (Unreleased)
 
+#### :boom: Breaking Change
+
+- Parse `assert` as a regular function. `assert` is no longer a unary expression. Example: before `assert 1 == 2` is parsed as `(assert 1) == 2`, now it is parsed as `assert(1 == 2)`. https://github.com/rescript-lang/rescript-compiler/pull/6180
+
 #### :bug: Bug Fix
+
+- Make "rescript format" work with node 10 again and set minimum required node version to 10 in package.json. https://github.com/rescript-lang/rescript-compiler/pull/6186
 - GenType: add support for custom `@tag` in variant type declaration. https://github.com/rescript-lang/rescript-compiler/pull/6137/files
 
 # 11.0.0-alpha.4