+++ CHANGELOG.md
@@ -17,8 +17,8 @@
 - Add support for Dynamic import. https://github.com/rescript-lang/rescript-compiler/pull/5703
 - GenType: Add `moduleResolution` option to customize extensions on emitted import statements. This helps to adjust output compatibility with TypeScript projects using ESM. https://github.com/rescript-lang/rescript-compiler/pull/6182
   - `node` (default): Drop extensions.
-  - `node16`: Use TS output's extensions (e.g. `.gen.js`). Make it ESM-compatible.
-  - `bundler`: Use TS input's extensions (e.g. `.gen.tsx`). Make it ESM-compatible.
+  - `node16`: Use TS output's extensions. Make it ESM-compatible.
+  - `bundler`: Use TS input's extensions. Make it ESM-compatible.
 
 #### :boom: Breaking Change