+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 > - :house: [Internal]
 > - :nail_care: [Polish]
 
-# 11.0.0-alpha.5 (Unreleased)
+# 11.0.0-alpha.5
 
 #### :rocket: Main New Feature